applied database and console log and el to element changes
