+++ pages/library/viewBook.js
@@ -1,6 +1,8 @@
 import Link from "next/link";
 import Image from "next/image";
 import styled from "styled-components";
+import { useEffect } from "react";
+import useSWR from "swr";
 
 const BookPageStyle = styled.ul`
   position: relative;
